requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.9.0=pypi_0
attrs=19.3.0=py_0
backcall=0.2.0=py_0
blas=1.0=mkl
bleach=3.1.5=py_0
ca-certificates=2020.6.24=0
cachetools=4.1.1=pypi_0
catalyst=20.6=pypi_0
certifi=2020.6.20=py38_0
chardet=3.0.4=pypi_0
click=7.1.2=pypi_0
crc32c=2.0.1=pypi_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=pypi_0
dbus=1.13.16=hb2f20db_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
deprecation=2.1.0=pypi_0
entrypoints=0.3=py38_0
expat=2.2.9=he6710b0_2
filelock=3.0.12=pypi_0
fontconfig=2.13.0=h9420a91_0
freetype=2.10.2=h5ab3b9f_0
future=0.18.2=pypi_0
gitdb=4.0.5=pypi_0
gitpython=3.1.3=pypi_0
glib=2.65.0=h3eb4bd4_0
google-auth=1.18.0=pypi_0
google-auth-oauthlib=0.4.1=pypi_0
grpcio=1.30.0=pypi_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb31296c_0
icu=58.2=he6710b0_3
idna=2.10=pypi_0
importlib-metadata=1.7.0=py38_0
importlib_metadata=1.7.0=0
intel-openmp=2020.1=217
ipykernel=5.3.0=py38h5ca1d4c_0
ipython=7.16.1=py38h5ca1d4c_0
ipython_genutils=0.2.0=py38_0
ipywidgets=7.5.1=py_0
jedi=0.17.1=py38_0
jinja2=2.11.2=py_0
joblib=0.16.0=pypi_0
jpeg=9b=h024ee3a_2
jsonschema=3.2.0=py38_0
jupyter=1.0.0=py38_7
jupyter_client=6.1.3=py_0
jupyter_console=6.1.0=py_0
jupyter_core=4.6.3=py38_0
kiwisolver=1.2.0=pypi_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h7b6447c_0
libffi=3.3=he6710b0_1
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=he19cac6_1
lz4-c=1.9.2=he6710b0_0
markdown=3.2.2=pypi_0
markupsafe=1.1.1=py38h7b6447c_0
matplotlib=3.2.2=pypi_0
mistune=0.8.4=py38h7b6447c_1000
mkl=2020.1=217
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.1.0=py38h23d657b_0
mkl_random=1.1.1=py38h0573a6f_0
nbconvert=5.6.1=py38_0
nbformat=5.0.7=py_0
ncurses=6.2=he6710b0_1
ninja=1.9.0=py38hfd86e86_0
nltk=3.5=pypi_0
notebook=6.0.3=py38_0
numpy=1.18.5=py38ha1c710e_0
numpy-base=1.18.5=py38hde5b4d6_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py_0
openssl=1.1.1g=h7b6447c_0
packaging=20.4=py_0
pandas=1.0.5=py38h0573a6f_0
pandoc=2.9.2.1=0
pandocfilters=1.4.2=py38_1
parso=0.7.0=py_0
pcre=8.44=he6710b0_0
pexpect=4.8.0=py38_0
pickleshare=0.7.5=py38_1000
pillow=7.1.2=py38hb39fc2d_0
pip=20.1.1=py38_1
plotly=4.8.2=pypi_0
prometheus_client=0.8.0=py_0
prompt-toolkit=3.0.5=py_0
prompt_toolkit=3.0.5=0
protobuf=3.12.2=pypi_0
ptyprocess=0.6.0=py38_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pygments=2.6.1=py_0
pyparsing=2.4.7=py_0
pyqt=5.9.2=py38h05f1152_4
pyrsistent=0.16.0=py38h7b6447c_0
python=3.8.3=hcff3b4d_2
python-dateutil=2.8.1=py_0
pytorch=1.5.1=py3.8_cuda10.2.89_cudnn7.6.5_0
pytorch-nlp=0.5.0=pypi_0
pytz=2020.1=py_0
pyyaml=5.3.1=pypi_0
pyzmq=19.0.1=py38he6710b0_1
qt=5.9.7=h5867ecd_1
qtconsole=4.7.5=py_0
qtpy=1.9.0=py_0
readline=8.0=h7b6447c_0
regex=2020.6.8=pypi_0
requests=2.24.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
retrying=1.3.3=pypi_0
rsa=4.6=pypi_0
sacremoses=0.0.43=pypi_0
scikit-learn=0.23.1=pypi_0
scipy=1.5.0=pypi_0
send2trash=1.5.0=py38_0
sentencepiece=0.1.91=pypi_0
setuptools=47.3.1=py38_0
sip=4.19.13=py38he6710b0_0
six=1.15.0=py_0
smmap=3.0.4=pypi_0
sqlite=3.32.3=h62c20be_0
tensorboard=2.2.2=pypi_0
tensorboard-plugin-wit=1.7.0=pypi_0
tensorboardx=2.0=pypi_0
terminado=0.8.3=py38_0
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pypi_0
tk=8.6.10=hbc83047_0
tokenizers=0.8.0rc4=pypi_0
torchvision=0.6.1=py38_cu102
tornado=6.0.4=py38h7b6447c_1
tqdm=4.47.0=py_0
traitlets=4.3.3=py38_0
transformers=3.0.1=pypi_0
urllib3=1.25.9=pypi_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py38_1
werkzeug=1.0.1=pypi_0
wheel=0.34.2=py38_0
widgetsnbextension=3.5.1=py38_0
xz=5.2.5=h7b6447c_0
zeromq=4.3.2=he6710b0_2
zipp=3.1.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.4=h0b5b093_3

jsonlines~=1.2.0
numpy~=1.18.5
pandas~=1.0.5
torchtext~=0.6.0
nltk~=3.5
tqdm~=4.47.0
catalyst~=20.6
scikit-learn~=0.23.1
transformers~=3.0.1